backend/requirements.txt
# Django and Core Framework
Django==4.2.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# Environment and Configuration
python-dotenv==1.0.0

# AI Integration
google-genai==0.3.2

# Date and Time Handling
python-dateutil==2.8.2

# HTTP Requests (for testing)
requests==2.31.0
# Development and Testing
pytest==7.4.0
pytest-django==4.5.2

# Code Quality
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Type Checking
mypy==1.5.1
django-stubs==4.2.7

# Production (optional)
gunicorn==21.2.0
whitenoise==6.5.0 